Remove unused models and related code from models2.py to streamline the application structure.
