+++ fastapi/app/database.py
@@ -2,6 +2,7 @@ import os
 from sqlmodel import create_engine, SQLModel
 from sqlalchemy.orm import sessionmaker
 
+from app import models
 
 DATABASE_URL = os.environ.get("DATABASE_URL")
 if not DATABASE_URL:
@@ -13,7 +14,6 @@ SessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=engine)
 
 def create_db_and_tables():
     # Import models so that SQLModel.metadata is populated before creating tables
-    from . import models  # noqa: F401
     SQLModel.metadata.create_all(bind=engine)
 
 def get_db():
